pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "incident-response-env"
version = "1.0.0"
description = "OpenEnv environment for AI-powered production incident response and on-call triage"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.10"
keywords = ["openenv", "reinforcement-learning", "incident-response", "sre", "devops", "agent"]

dependencies = [
    "openenv-core>=0.0.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]
baseline = [
    "openai>=1.0.0",
]

[project.scripts]
# Convenience entry point to start the server
incident-response-env = "server.app:main"

[tool.hatch.build.targets.wheel]
packages = ["server"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py310"
